backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.5
Pillow==10.1.0
